requirements.txt
Flask
Flask-Cors
PyJWT
bcrypt
pymongo
python-dotenv
requests
pytest
mongomock
